Started Electron integration.
