+++ client/index.html
@@ -13,5 +13,5 @@
   <body>
     <div id="root"></div>
   </body>
-  <script src="./bundle.js"></script>
+  <script src="client/dist/bundle.js"></script>
 </html>
