--- client/main.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { render } from 'react-dom';
-import './styles/main.css';
-import App from './App';
-
-render(<App />, document.getElementById('root'));
