+++ package.json
@@ -8,6 +8,7 @@
     "@emotion/styled": "^11.3.0",
     "@mui/icons-material": "^5.0.4",
     "@mui/material": "^5.0.4",
+    "electron": "^15.3.0",
     "express": "^4.17.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -34,8 +35,10 @@
     "webpack-hot-middleware": "^2.24.3"
   },
   "scripts": {
-    "start": "nodemon ./server/server.js",
+    "start": "node ./server/server.js",
+    "build": "webpack",
     "dev": "concurrently \"webpack serve --hot\" \"nodemon ./server/server.js\"",
+    "e": "electron .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
